requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements.txt requirements.in
#
aiohttp==3.5.4
aioredis==1.2.0
async-timeout==3.0.1      # via aiohttp, aioredis
asyncpg==0.18.3
attrs==19.1.0             # via aiohttp
certifi==2019.6.16        # via sentry-sdk
chardet==3.0.4            # via aiohttp
hiredis==1.0.0            # via aioredis
idna==2.8                 # via yarl
git+git://github.com/IOMirea/rpc@v0.2.0#egg=iomirea_rpc
multidict==4.5.2          # via aiohttp, yarl
pyyaml==5.1.1
sentry-sdk==0.11.2
urllib3==1.25.3           # via sentry-sdk
uvloop==0.13.0
yarl==1.3.0               # via aiohttp
